Fixed: Added author_questions

+++ app/controllers/questions_controller.rb
@@ -6,6 +6,8 @@ class QuestionsController < ApplicationController
   expose :answer, -> { Answer.new }
 
   def create
+    @exposed_question = current_user.author_questions.new(question_params)
+
     if question.save
       redirect_to question_path(question), notice: 'Your question successfully created.'
     else
@@ -22,8 +24,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question was successfully deleted.'
+    else
+      redirect_to questions_path, notice: 'You have no rights to do this.'
+    end
   end
 
   private
